chore: rename polyfills to js-polyfills
