+++ package.json
@@ -114,7 +114,7 @@
     "@react-native-community/cli-platform-ios": "10.0.0-alpha.3",
     "@react-native/assets": "1.0.0",
     "@react-native/normalize-color": "2.1.0",
-    "@react-native/polyfills": "2.0.0",
+    "@react-native/polyfills": "^0.72.0",
     "abort-controller": "^3.0.0",
     "anser": "^1.4.9",
     "base64-js": "^1.1.2",
