+++ packages/js-polyfills/BUCK
@@ -21,8 +21,8 @@ yarn_workspace(
 )
 
 rn_library(
-    name = "polyfills",
-    base_path = relative_path_to_js_root() + "node_modules/@react-native/polyfills/",
+    name = "js-polyfills",
+    base_path = relative_path_to_js_root() + "node_modules/@react-native/js-polyfills/",
     is_polyfill = True,
     labels = [
         "pfh:ReactNative_CommonInfrastructurePlaceholder",
