+++ packages/js-polyfills/error-guard.js
