+++ packages/js-polyfills/package.json
@@ -1,11 +1,11 @@
 {
-  "name": "@react-native/polyfills",
-  "version": "2.0.0",
+  "name": "@react-native/js-polyfills",
+  "version": "0.71.0",
   "description": "Polyfills for React Native.",
   "repository": {
     "type": "git",
     "url": "git@github.com:facebook/react-native.git",
-    "directory": "packages/polyfills"
+    "directory": "packages/js-polyfills"
   },
   "license": "MIT"
 }
